microservices/farm_service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
motor>=3.3.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pyjwt>=2.8.0
httpx>=0.26.0
shapely>=2.0.0
